+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
- 
+
+## [Unreleased]
+
+#### Breaking Changes
+
+- Using `*increments()` for a primary key definition now requires adding `.primary()` to the column definition. `id()` keeps creating the primary key automatically.
+
+### Fixed
+
+- `*increments()` could not be used for non primary key columns
+- Schema honors explicitly provided connection details instead of always reloading them from the config file
+
 ## [3.0.0] - 2025-03-30
    
 ### Changed
@@ -16,8 +27,9 @@
   
 ### Fixed
 
+- QueryBuilder requires `database.py` even when passing connection detain in directly
+- Several classes would import DB (ConnectionResolver) from config even if inline connection details are provided   
  - Model `update` and `delete` not casting passed values
-
 
 ## [2.24.0] - 2025-01-23